Split test_hooks file into test file per hook.
