+++ zerver/tests/webhooks/test_IFTTT.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+from zerver.lib.test_helpers import WebhookTestCase
+
+class IFTTTHookTests(WebhookTestCase):
+    STREAM_NAME = 'ifttt'
+    URL_TEMPLATE = "/api/v1/external/ifttt?stream={stream}&api_key={api_key}"
+    FIXTURE_DIR_NAME = 'ifttt'
+
+    def test_ifttt_when_subject_and_body_are_correct(self):
+        # type: () -> None
+        expected_subject = u"Email sent from email@email.com"
+        expected_message = u"Email subject: Subject"
+        self.send_and_test_stream_message('correct_subject_and_body', expected_subject, expected_message)
